more SelectRoomViewController stuff

+++ ota-ios/ota-ios/Controllers/BookViewController.m
@@ -34,7 +34,7 @@
 - (void)requestFinished:(NSData *)responseData {
     NSString *respString = [[NSString alloc] initWithData:responseData encoding:NSUTF8StringEncoding];
     EanHotelRoomReservationResponse *hrrr = [EanHotelRoomReservationResponse roomReservationFromData:responseData];
-    self.itineraryOutlet.text = [NSString stringWithFormat:@"Itin:%ld", hrrr.itineraryId];
+    self.itineraryOutlet.text = [NSString stringWithFormat:@"Itin:%ld", (long)hrrr.itineraryId];
     NSLog(@"%@.%@:::%@", self.class, NSStringFromSelector(_cmd), respString);
 }
